pyproject.toml
[tool.poetry]
name = "allure-api-client"
version = "1.2.0"
description = "API Client for Python ^3.11 and Pytest projects. It includes the Allure request/response hooks and the status code checking inside the main method. Request hook generates cURL, Response hook represents the useful information."
authors = ["amaym0n <loco.skitls@gmail.com>"]
license = "MIT License"
readme = "README.md"
packages = [
    { include = "allure_api_client" },
]

[tool.poetry.dependencies]
python = "^3.11"
httpx = "^0.25.2"
httpx-auth = "^0.18.0"
allure-pytest = "^2.13.2"
pydantic = "^2.5.3"

[tool.poetry.group.dev.dependencies]
poetry = "^1.7.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
